test_requirements.txt
nose
should_dsl
noseOfYeti>=1.4.7
fudge
django_nose
pinocchio
